core/Solve: use res.SingleRef
* use the result's st for the FS
* use the original input st for the image config
* split Image code out of filesystem.go and into image.go
